apiфывфы

+++ upload_files.py
@@ -36,17 +36,18 @@ async def upload_statistic_files_async(order=None):
     """Отправка на я.диск файла с ненайденными артикулами"""
     current_directory = os.getcwd()
     directory = os.path.join(current_directory, 'Файлы связанные с заказом')
-    tasks = []
+    if os.path.exists(directory) and len(os.listdir(directory)) > 0:
+        tasks = []
 
-    async with aiohttp.ClientSession() as session:
-        for file in os.listdir(directory):
-            if order in file:
-                file_path = os.path.join(directory, file)
-                destination_path = f"/Отчеты/{file}"
-                task = upload_file(session, file_path, destination_path)
-                tasks.append(task)
+        async with aiohttp.ClientSession() as session:
+            for file in os.listdir(directory):
+                if order in file:
+                    file_path = os.path.join(directory, file)
+                    destination_path = f"/Отчеты/{file}"
+                    task = upload_file(session, file_path, destination_path)
+                    tasks.append(task)
 
-        await asyncio.gather(*tasks)
+            await asyncio.gather(*tasks)
 
 
 if __name__ == '__main__':
